requirements.txt
pyyaml
structlog
simple-plugin-loader
asyncudp
pyserial
